bug fixed
bug fixed

+++ School/School/Form1.cs
@@ -113,9 +113,13 @@ namespace School
                     int group = Convert.ToInt32(textBox3.Text);
                     string path3 = "gruppe.txt";
                     File.WriteAllText(path3, Convert.ToString(group));
-                    string psw = textBox2.Text;
                     string path = "psw.txt";
-                    File.WriteAllText(path, psw);
+                    if (!File.Exists(path))
+                    {
+                        string psw = textBox2.Text;
+                        File.WriteAllText(path, psw);
+                    }
+
                     MessageBox.Show("Login erfolgreich", ".");
                     Application.Restart();
                 }
